Remove mediator pattern & tidy up code
I realised that I would have to rewrite a lot of the code to make the
mediator pattern work, because I didn't have the pattern in mind when I
wrote the core code in toDoList.js. So I created a separate branch in
git and fixed the code to work without this pattern.